Omaha #3049 - entire data directory cannot be removed because it is mounted
The script does still verify that the variable reference to the data directory is set. Thus, the root
directory could only be removed as a result of malicious intent. A user could intentionally not
set the data directory and remove the verification check or the user could intentionally set the
data directory to the root location.


Former-commit-id: 604c5a4d10e2529d39f982efa14cf9bbfe7a3867 [formerly 429aec5855f6f73655b1239aaa03bc1d0ad99aa4 [formerly 6a5deba786d9360cf1f6c6396ffcf64aef752de9] [formerly 604c5a4d10e2529d39f982efa14cf9bbfe7a3867 [formerly efe867dc134027eb7cdf393473476fcc4ec0f6f8]]]
Former-commit-id: 429aec5855f6f73655b1239aaa03bc1d0ad99aa4 [formerly 6a5deba786d9360cf1f6c6396ffcf64aef752de9]
Former-commit-id: 429aec5855f6f73655b1239aaa03bc1d0ad99aa4
Former-commit-id: 9787e4f1d84f186abc9b4564124a74dd7688fc80