Create application-0b7df094bf.js
This commit is intended to get the search functionality working on the website again.  This is not the ideal solution -- ideally the mkdocs-unidata theme will be updated and distributed with this change so that the gh-action works with that latest version and the search will work properly.  This file should be removed when that "proper" change has been made and mkdocs-unidata theme has been updated on pip.