Two small changes to the edex installation webpage:
Added the gcc-c++ dependency to the edex install page.  Note: this can be removed when we put out a new version of the installation rpms because this will be included as a dependency.

Moved the `edex setup` command step 2 to step 3 and add that the user should be running as user awips first, and to use sudo to run the edex setup command.