Updated McIdas files to properly ingest and removed legacy satellite menus that are no longer used.
While the composite imagery (mcidas) data being sent via UNIWISC includes the new GOES 16 and 17 imagery, the areaNames.xml and mcidas plugin is not set up to "read" and store the different sectors. Because all of the GOES 16 and 17 single channel data is available for each sector already it did not seem necessary at this time to also have the mcidas imagery. For simplicity the mcidas composite imagery for the new GOES was removed. The Global, Arctic and Antarctic imagery remain available here. I did have to comment out some of the data in the areaNames.xml so the sector (AREA####) could be stored correctly for global.

The legacy GOES West (GOES-15) submenu was removed as we are no longer receiving the data.

The legacy sounder imagery submenu was removed as we are no longer receiving the legacy data. GOES 16 derived products can be found in the GOES menus.

The legacy satellite winds got consolodated to only the ASACT winds as we are no longer receiving legacy satellite wind data. GOES 16/17 satellite derived winds can be found in the GOES menus.