Updated the units for Absolute Vorticity in parameters.xml from 1/s to /s.
The previous check-in of this file, the units were updated to 1/s which was causing issues in CAVE displaying. For the native pressure heights (000, 850, 700, 500, 250 mb) the units were wrong and it wasn't displaying any data, but the interpolated heights were showing up correct. I also noticed this when running python-awips and getting the "unit" of AV as "count" instead of /s.

On systems where this is already installed, I had to manually update the parameter database table:
   update parameter set unit = '/s' where abbreviation like 'AV';